Asset and Mesh pipeline improvements + much more - Asset Registry is now much more robust, especially with file paths - Cleaned up Asset Metadata and its uses - Added jump flood outline effect for selected objects - .vam renamed to .vmesh - Fixed all scenes + asset registry - Removed a lot of unused asset garbage - Added Asset Manager UI panel with asset registory information - Added a lot of UI feedback for various things such as trying to use invalid assets - Content Browser panel now has Show in Explorer and Open Externally options - Sky light intensity now works - Sky light dynamic sky settings are now serialized

+++ Engine/Source/Editor/DefaultAssetEditors.cpp
@@ -22,8 +22,8 @@ namespace Vanta {
 		if (!m_Asset)
 			SetOpen(false);
 
-		float textureWidth = m_Asset->GetWidth();
-		float textureHeight = m_Asset->GetHeight();
+		float textureWidth =  (float)m_Asset->GetWidth();
+		float textureHeight =  (float)m_Asset->GetHeight();
 		//float bitsPerPixel = Texture::GetBPP(m_Asset->GetFormat());
 		float imageSize = ImGui::GetWindowWidth() - 40;
 		imageSize = glm::min(imageSize, 500.0f);
